Moved emails() from Arbitraries to ArbitrariesFacadeImpl, added EdgeCasesMode.NONE to tests, removed unnecessary i
